% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnomalyDetectionMethods.R
\docType{package}
\name{AnomalyDetectionMethods-package}
\alias{AnomalyDetectionMethods}
\alias{AnomalyDetectionMethods-package}
\title{AnomalyDetectionMethods: Outlier Detection for Value-Frequency Data}
\description{
Provides cross-sectional outlier detection methods and visualization utilities
for numeric value-frequency distributions.
}
\author{
\strong{Maintainer}: Anthony Sena \email{sena@ohdsi.org}

}
\keyword{internal}
