% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_reporting.R
\name{plotAnomalyFeatureScatter}
\alias{plotAnomalyFeatureScatter}
\title{Plot anomaly results in feature space}
\usage{
plotAnomalyFeatureScatter(
  df,
  xColumn,
  yColumn,
  frequencyColumn = "frequency",
  scoreColumn = NULL,
  flagColumn = "isOutlier"
)
}
\arguments{
\item{df}{A data.frame containing anomaly results.}

\item{xColumn}{Name of the feature column for the x-axis.}

\item{yColumn}{Name of the feature column for the y-axis.}

\item{frequencyColumn}{Name of the frequency column used for point size.}

\item{scoreColumn}{Optional name of a numeric score column used for color.}

\item{flagColumn}{Name of the logical outlier flag column used for color
when `scoreColumn` is `NULL`.}
}
\value{
A ggplot object.
}
\description{
Creates a two-feature scatter plot with point size mapped to `frequency` and
optional color mapping to a score column or outlier flag.
}
